Move filters into own package
